Support multiple IP addresses from hostname -I
If docker is running multiple IP addresses are returned by hostname -I,
we take the first one now.

Also if we update the IP address and it's shorter (in characters) than
the last IP address we have a corrupt wsl line, we now add a couple
empty lines to try and mitigate that.